[server][dvc] Lag measurement safeguard and logging improvements
It seems that sometimes the offset lag used in lag measurement is -1,
which results in the lag being considered acceptable. This commit
adds a safeguard to prevent prematurely declaring ready-to-serve, and
to log a stacktrace in cases where the offset lag is negative.

There are also various other log tweaks to make logs less voluminous
and more clear.