[dvc][server] Logging tweaks for PartitionTracker
Added more details when logging about unregistered producers.

Also substituted "leader metadata" for "producer" in the logging of
DIV errors, since the leader metadata is unfortunately no longer
used exclusively by leaders. Left some comments in the code to that
effect, about potentially needing to refactor/fix this, but did not
introduce such change yet.

Also introduced new utility functions in CollectionUtils to test if
a map or collection is null or empty.