[controller] VeniceControllerStateModel race condition fix
In the STANDBY to LEADER transition, there is a check for whether the
helixManager is initialized, but not for whether the clusterResources
are initialized. The problem is that when calling reset(), these two
resources are torn down in the order of the clusterResources first
and the helixManager second, which means that in the aforementioned
state transition, it is possible for the clusterResources to already
be null while the helixManager is still initialized. This could
result in these resources not getting properly reinitialized.