[controller] Deprecated the Helix Message Channel
The Helix Message Channel is a legacy mechanism for letting the controller
send kill job commands to the servers. It has been replaced a long time ago
by the Participant Store but the config defaults did not reflect that.

This commit aligns the default config with production best practices. A
later commit will delete this code entirely.