[controller] Deprecated the Helix Message Channel and T-Wise Shared Consumer
The Helix Message Channel is a legacy mechanism for letting the controller
send kill job commands to the servers. It has been replaced a long time ago
by the Participant Store but the config defaults did not reflect that.

The Topic-Wise Shared Consumer is also a legacy mode, which has since been
replaced in production by the Partition-Wise Shared consumer.

This commit aligns the defaults for both of these configs with production
best practices. A later commit will delete this code entirely.

Miscellaneous:

- Removed some unused partition params in various SIT functions.

- Skipped the TransientRecordCache lookup in more situations.

- Reduced logging verbosity in HelixSchemaAccessor.

- In a system store creation code path, made use of default partition count
  rather than failing.